+++ index.html
@@ -8,6 +8,7 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
   <script src="main.js"></script>
+  <script type="module" src="main.js"></script>
 </head>
 
 <body>
