+++ util.js
@@ -1,0 +1,12 @@
+function getRandomInt(a, b) {
+  const min = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const max = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+function checkStringLenght(str, len) {
+  return str.length <= len;
+}
+
+export { getRandomInt, checkStringLenght };
